Fixed file paths after reorganization
