+++ src/GameContext/useGameById.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import {API_RAWG} from './apiConfig';
+import { API_RAWG } from '../services/api';
 
 function useGameById(id){
 	const [gameById, setGame] = useState([]);
