+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.css';
+import App from './components/App';
 
-import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
